Se agregan los thunder y validaciones

+++ Opinion Manager/src/comment/comment.controller.js
@@ -25,18 +25,17 @@ export const update = async (req,res) =>{
     try{
         let { id } = req.params
         let data = req.body
-        let user = await User.findOne({_id: data.user })
-        if(!user) return res.status(404).send({message: 'User not found'})
-        let post = await Post.findOne({_id: data.post})
-        if(!post) return res.status(404).send({message: 'Post not found'})
+        let uid = req.user._id
         let update = checkUpdatePost(data, id)
+        //let comment = await Comment.findOne({_id: id, user: uid})
+        //if(!comment) return res.status(400).send({message: 'Comment not found'})
         if(!update) return res.status(400).send({message: 'Have submitted some data that cannot be updated or missing data'})
         let updatedComment = await Post.findOneAndUpdate(
             {_id: id},
             data, 
             {new: true}
         )
-        if(!updatedComment) return res.status(401).send({message: 'Comment not found and not updated'})
+        //if(!updatedComment) return res.status(401).send({message: 'Comment not found and not updated'})
         return res.send({message: 'Updated comment', updatedComment})
     }catch(err){
         console.error(err)
@@ -48,9 +47,14 @@ export const update = async (req,res) =>{
 export const deleteU = async (req,res)=>{
     try{ 
         let { id } = req.params
-        let deletedComment = await Comment.findOneAndDelete({_id: id})
-        if(!deletedComment) return res.status(404).send({message: 'Comment not found and not deleted'})
-        return res.send({message: `Comment with tittle ${deletedComment.tittle} deleted successfully`})
+        let uid = req.user._id
+
+        let comment = await Comment.findOne({_id: id, user: uid})
+        if(!comment)
+        return res.status(404).send({message: 'Comment not found and not deleted'})
+        let deletedComment = await Comment.findOneAndDelete({_id: id, user: uid})
+        if(!deletedComment) return res.status(404).send({message: 'Error deleting comment'})
+        return res.send({message: `Comment deleted successfully`})
     }catch(err){
         console.error(err)
         return res.status(500).send({message: 'Error deleting comment'})
